requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
cloudinary
python-dotenv